+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.0.45] - 2023-06-15
+### Added
+* slight optimization to `ListPool`
+* optional parameters to `IPoolable`
+* optional parameter handling to `PoolSystem`
+### Changed
+* `CursorManager.Update` to `CursorManager.LateUpdate`
+
 ## [0.0.44] - 2023-04-12
 ### Added
 * `PoolSystem` and `IPoolable<T>` which serves as a similar framework to `ListPool`